+++ db/migrate/20190117135150_add_voucher_to_project_item.rb
@@ -1,0 +1,5 @@
+class AddVoucherToProjectItem < ActiveRecord::Migration[5.2]
+  def change
+    add_column :project_items, :voucher_id, :string, default: "", null: false
+  end
+end