+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ### Added
 - User needs to select "Research area" while ordering service (@mkasztelnik)
+- Add voucher support in service ordering (@michal-szostak)
 
 ### Changed